requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
openai==1.3.5
anthropic==0.34.0
requests==2.31.0
python-multipart==0.0.6
jinja2==3.1.2
aiofiles==23.2.1
python-dotenv==1.0.0
# Testing dependencies
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
# Code quality tools
black==25.9.0
flake8==7.3.0
isort==7.0.0